Move types for more digestible file
